# openspec/specifications/apis/openapi.yaml
openapi: 3.0.3
info:
  title: AI Tool Usage Tracker API
  version: 1.0.0
  description: |
    REST API for the AI Tool Usage Tracker platform (Phase 1 MVP).

    **Authentication:** JWT Bearer token except `/auth/login`, `/auth/refresh`, and `/health`.

    **Errors:** RFC 7807 Problem Details (`application/problem+json`).

    **Traceability:** FR-PLT-001, ADR-012, NFR-SEC-003/004/006.

    ## RBAC Roles and Access

    | Role | Scope |
    |---|---|
    | `super_admin` | Full access to all endpoints and data across the entire organisation |
    | `team_admin` | Access to members, tools (read), teams (read), thresholds, uploads, collectors (read), and dashboard analytics — all scoped to managed teams only |
    | `finance_viewer` | Access to reports (read) and `GET /dashboard/my-usage` only |
    | `auditor` | Access to audit logs and `GET /dashboard/my-usage` only |
    | `team_member` | Access to `GET /dashboard/my-usage` only |
  contact:
    name: Platform Team
  license:
    name: Proprietary
servers:
  - url: https://api.ai-tool-tracker.example/api/v1
    description: Production
  - url: https://staging-api.ai-tool-tracker.example/api/v1
    description: Staging
  - url: http://localhost:8000/api/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /credentials:
    get:
      tags: [Credentials]
      summary: List API credentials
      description: Returns masked credentials only. **Roles:** `super_admin`.
      operationId: listCredentials
      parameters:
        - $ref: './components/parameters.yaml#/ToolId'
        - $ref: './components/parameters.yaml#/TeamId'
        - $ref: './components/parameters.yaml#/Limit'
        - $ref: './components/parameters.yaml#/Cursor'
      responses:
        '200':
          description: Credential list (masked)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: './components/schemas.yaml#/Credential'
                  meta:
                    $ref: './components/schemas.yaml#/PaginationMeta'
        '403':
          $ref: './components/responses.yaml#/Forbidden'
    post:
      tags: [Credentials]
      summary: Create API credential
      description: |
        **Roles:** `super_admin`.

        **Validation:** `secret_value` write-only; stored AES-256 encrypted (NFR-SEC-001).
        Full secret never returned after creation.
      operationId: createCredential
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/schemas.yaml#/CredentialCreateRequest'
            examples:
              default:
                $ref: './components/examples.yaml#/CredentialCreateExample'
      responses:
        '201':
          description: Credential created (masked)
          content:
            application/json:
              schema:
                $ref: './components/schemas.yaml#/Credential'
        '403':
          $ref: './components/responses.yaml#/Forbidden'
        '422':
          $ref: './components/responses.yaml#/UnprocessableEntity'
...
  /credentials/{credentialId}:
    delete:
      tags: [Credentials]
      summary: Delete API credential
      operationId: deleteCredential
      parameters:
        - $ref: './components/parameters.yaml#/CredentialId'
      responses:
        '204':
          description: Credential deleted
        '403':
          $ref: './components/responses.yaml#/Forbidden'
        '404':
          $ref: './components/responses.yaml#/NotFound'
...
  /audit-logs:
    get:
      tags: [Audit]
      summary: Query audit logs
      description: '**Roles:** `super_admin`, `auditor`. Append-only log (NFR-AUD-001).'
      operationId: listAuditLogs
      parameters:
        - $ref: './components/parameters.yaml#/DateFrom'
        - $ref: './components/parameters.yaml#/DateTo'
        - name: actor_id
          in: query
          schema:
            type: string
            format: uuid
        - name: action
          in: query
          schema:
            type: string
        - name: resource_type
          in: query
          schema:
            type: string
        - $ref: './components/parameters.yaml#/Limit'
        - $ref: './components/parameters.yaml#/Cursor'
      responses:
        '200':
          description: Audit log entries
          content:
            application/json:
              schema:
                $ref: './components/schemas.yaml#/AuditLogListResponse'
              examples:
                default:
                  value:
                    data:
                      - id: cc0e8400-e29b-41d4-a716-446655440060
                        actor_id: 550e8400-e29b-41d4-a716-446655440001
                        actor_email: admin@acme.example
                        action: credential.rotate
                        resource_type: credential
                        resource_id: dd0e8400-e29b-41d4-a716-446655440070
                        outcome: success
                        source_ip: 203.0.113.10
                        correlation_id: f47ac10b-58cc-4372-a567-0e02b2c3d479
                        created_at: '2026-06-10T09:30:00Z'
                    meta:
                      limit: 20
                      has_more: false
                      next_cursor: null
        '403':
          $ref: './components/responses.yaml#/Forbidden'
  /audit-logs/export:
    post:
      tags: [Audit]
      summary: Export audit logs to CSV
      description: '**Roles:** `super_admin`, `auditor`.'
      operationId: exportAuditLogs
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [from, to]
              properties:
                from:
                  type: string
                  format: date-time
                to:
                  type: string
                  format: date-time
                actor_id:
                  type: string
                  format: uuid
                action:
                  type: string
      responses:
        '200':
          description: CSV export
          content:
            text/csv:
              schema:
                type: string
                format: binary
        '403':
          $ref: './components/responses.yaml#/Forbidden'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT from POST /auth/login. Required on all endpoints except /auth/login, /auth/refresh, /health.